requirements.txt
Django===4.0
psycopg2-binary===2.9.6
django-filter===23.2 
djangorestframework===3.14.0
drf-nested-routers===0.93.4